move network types to its own package
