+++ pkg/domain/link.go
@@ -6,6 +6,7 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"fmt"
+	"github.com/skupperproject/skupper/pkg/network"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -27,7 +28,7 @@ type LinkHandler interface {
 	StatusAll() ([]types.LinkStatus, error)
 	Status(name string) (types.LinkStatus, error)
 	Detail(link types.LinkStatus) (map[string]string, error)
-	RemoteLinks(ctx context.Context) ([]*types.RemoteLinkInfo, error)
+	RemoteLinks(ctx context.Context) ([]*network.RemoteLinkInfo, error)
 }
 
 type SecretUpdateFn func(claim *corev1.Secret) error
